refactor(online-class): add schema file
