+++ packages/2024-online-class-prototype/app/_components/course-list.tsx
@@ -1,5 +1,5 @@
 import { z } from 'zod'
-import { courseObject } from '../page'
+import { courseObject } from './schema'
 import CourseItem from './course-item'
 
 type Props = {
