+++ packages/2024-online-class-prototype/app/_components/payment-flow.tsx
@@ -1,5 +1,5 @@
 import { z } from 'zod'
-import { dataSchema } from '../page'
+import { dataSchema } from './schema'
 import CustomImage from '@/components/custom-image'
 
 type Props = {
